# fix DSS performance