@charleskawczynski 2022-04-29
Looks like this did it, awesome!

Main:
```
BenchmarkTools.Trial: 14 samples with 1 evaluation.
--
  | Range (min … max):  349.499 ms … 379.547 ms  ┊ GC (min … max): 0.00% … 5.44%
  | Time  (median):     374.925 ms               ┊ GC (median):    5.36%
  | Time  (mean ± σ):   371.951 ms ±   9.121 ms  ┊ GC (mean ± σ):  4.26% ± 2.29%
  |  
  | █     ▃
  | ▇▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▁▁▇▁▁▁▇▁▁▁▁█▁▁▇▁▇█▁▇▇ ▁
  | 349 ms           Histogram: frequency by time          380 ms <
  |  
  | Memory estimate: 98.63 MiB, allocs estimate: 1974780.
```

This PR:
```
BenchmarkTools.Trial: 24 samples with 1 evaluation.
--
  | Range (min … max):  213.333 ms … 221.598 ms  ┊ GC (min … max): 0.00% … 0.00%
  | Time  (median):     214.456 ms               ┊ GC (median):    0.00%
  | Time  (mean ± σ):   214.824 ms ±   1.632 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%
  |  
  | ▁ ▁ ▁▁█▁
  | ▆▁█▆█▆████▁▆▆▁▆▁▆▁▁▆▁▁▁▁▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆ ▁
  | 213 ms           Histogram: frequency by time          222 ms <
  |  
  | Memory estimate: 1.50 MiB, allocs estimate: 14844.
```