@charleskawczynski 2022-04-29
We should probably create a new release after this since it does result in a decent speedup